Renamed freeze -> bake
